+++ app/src/main/res/layout/fragment_create_update_habit.xml
@@ -6,40 +6,40 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".CreateUpdateHabitFragment">
+    tools:context=".ui.habits.CreateUpdateHabitFragment">
 
     <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/input_layout_name"
+        android:id="@+id/name_layout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:hint="습관 이름 (필수)"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
+        app:boxBackgroundMode="outline"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent">
 
         <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/edit_text_habit_name"
+            android:id="@+id/name"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:inputType="textCapSentences"
             android:maxLength="50"
-            tools:text="아침 조깅" />
+            tools:text="습관1" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/input_layout_description"
+        android:id="@+id/description_layout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
-        android:hint="상세 설명 (선택 사항)"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        app:layout_constraintTop_toBottomOf="@id/input_layout_name"
+        android:hint="상세 설명 (선택)"
+        app:boxBackgroundMode="outline"
+        app:layout_constraintTop_toBottomOf="@id/name_layout"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent">
 
         <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/edit_text_habit_description"
+            android:id="@+id/description"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:minLines="3"
@@ -47,7 +47,7 @@
             android:gravity="top|start"
             android:inputType="textMultiLine"
             android:scrollbars="vertical"
-            tools:text="매일 아침 7시에 집 근처 공원 2바퀴 돌기" />
+            tools:text="습관 상세 설명1" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <LinearLayout
@@ -56,12 +56,12 @@
         android:orientation="horizontal"
         android:gravity="end"
         android:layout_marginTop="24dp"
-        app:layout_constraintTop_toBottomOf="@id/input_layout_description"
+        app:layout_constraintTop_toBottomOf="@id/description_layout"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent">
 
         <Button
-            android:id="@+id/button_cancel"
+            android:id="@+id/cancel_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="취소"
@@ -69,11 +69,11 @@
             style="@style/Widget.MaterialComponents.Button.TextButton" />
 
         <Button
-            android:id="@+id/button_save"
+            android:id="@+id/save_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="저장하기"
-            android:enabled="false" />
+            android:enabled="true" />
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
