+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.HabitCheck" parent="Theme.AppCompat.DayNight.NoActionBar" />
+    <style name="Theme.HabitCheck" parent="Theme.Material3.DayNight.NoActionBar" />
 
 </resources>
